Leave the cache state as we found it
In order to avoid persisting DataFrames unnecessarily or unpersisting
DataFrames that users wanted to touch interactively after calling
analyze(), try to leave the cache state of DataFrames the way we found
it when analyze() was invoked.

`is_cached` is an undocumented DataFrame property, so be defensive about
accessing it in case a future Spark version removes it.