Add the `contrib` namespace, and two util fns.
The contrib namespace is a place for code that is being re-used
across multiple reports/investigations, and may or may not be ready
for others to start using it yet. One day this code dreams of being
promoted out of contrib and into the main package. Until then,
others are welcome to look at and even use your code, but should
not be surprised when breaking changes are silently made. If they
want reliability, then they should ask for/help with its promotion
out of contrib.