First step in a new direction.
This commit removes a lot of the existing code, and takes the first step on a new roadmap that takes mozanalysis beyond the `experiments` table, and aims to make safe and rigorous analyses easier than the alternative, while maintaining 'enough' flexibility.

This commit introduces the basic query infrastructure - most of the code that builds Spark queries.

Still to come in mozanalysis or mozreport (out of scope for this PR):
- A variety of stats functions to analyze the data output by `Experiment.get_per_client_data()`
- A library of pre-defined and blessed metrics
- Convenience functions for querying and analyzing time series (time since enrollment)
- Functions to visualize/present the output of stats functions