Work around bug in Spark joining.
If you start with a DataFrame, aggregate over it, then join from the aggregation to the original DataFrame, then Spark can raise an AnalysisException, complaining about duplicated names even if you've been careful to rename them. Specifically, this happened with 'submission_date_s3'.

Since it _actually is a unique name_, we can use `F.col('submission_date_s3')` instead of `data_source.submission_date_s3` to work around this problem. If someone supplies an `enrollments` table that _does_ have a `submission_date_s3` column then Spark justifiably raises a meaningful error.

Added the fix, a short comment explaining it with a reference to the bug[1], and a test that failed before the bug was fixed.

[1] https://issues.apache.org/jira/browse/SPARK-10925