Work around unsafe fork in pyspark
Touching core Objective C classes from a child process (in the case of
fork() without exec()) causes a defensive crash in recent versions of
macOS.

Pyspark forks unsafely, which triggers the defensive crash, which causes
tests to fail.

This doesn't fix anything -- it just disables the airbag -- but a proper
fix would have to land in pyspark.

Closes #40.