Transpose the result DataFrame for batch stats.
Also add documentation and tests for survival func code, and make a few other tweaks.

The result DataFrame was transposed so that for the survival func code, and for the future time series code, plotting should be easier. E.g. if we want to plot the expected uplift vs the threshold (or time) then it's one column in the resulting dataframe, and each error bar is another column.