Address review comments; use empirical bootstrap.
The biggest change here is that when bootstrapping for a single branch, we use the empirical/basic bootstrap rather than the percentile bootstrap, which is reputedly less biased for asymmetric distributions (see #5).

Interestingly, on our test suite's example asymmetric distribution, it agrees with the `bayesian_binary` method _less_ than the percentile bootstrap and I had to relax the error tolerance.

When comparing branches (i.e. computing uplifts), we still make the percentile-bootstrap-like approximation that the bootstrap samples act as samples from an appropriate posterior distribution.