Allow dependency injection for get_enrollments
`get_enrollments` needs to be able to reference a table of all
enrollment events. We need to be able to shortcut this in the tests, and
we'd like Experiment to be immutable, which means no monkeypatching.

Instead, allow the function that fetches the view to be passed in,
and call any callables that we receive as `study_type` arguments.

Further, align the APIs for the view-fetchers, and perform addon-version
filtering in the body of `get_enrollments` if it's required.