Add stats functions (#44)
Add stats functions.

Introduces stats functionality to mozanalysis. There is a module for Frequentist stats and a module for Bayesian stats. So far, most of the stats functions return credible intervals or confidence intervals over some sampled quantity.

The goal here is to strike the right balance between consistency and flexibility: we want the stats functions to be as interchangable as is sensible. They all accept data in the same format (the format output by `get_per_client_data`) and try to output data in as similar formats as is sensible. I hope this format is flexible enough to accommodate future functionality too, as people add their favourite stats methods.

Fixes #12, Fixes #14