Refactor time limits code.
Our next priorities are to introduce a metrics library and support for querying time series of metrics data. Both of these things add complexity to the queries that get sent to Spark, and the code that builds them. Here we lay some groundwork for those features by disentangling some logic from the `Experiment` class.

The main change here is the introduction of the `TimeLimits` class: an instance of TimeLimits holds all the temporal information used when querying data for an `Experiment`. It contains the start and end of the enrollment period, the details about the analysis window, and the relevant date range for the data source.

Previously, we tried to hold minimal state and compute quantities on the fly. Instead, `TimeLimits` precomputes and stores all these variables, which is conceptually easier (particularly when we start specifying a different set of variables when working with time series). We ensure self-consistency by providing a constructor with a near-minimal set of arguments, and by using validators.

This commit is not intended to be merged to master in its current state: it offers no new features or bug fixes. It is intended to form a base for future work on time series and metrics.

This commit includes a little cruft too: there is a bunch of commented-out WIP/prototype code to handle time series and a metrics library. I figured it was easier to include it here, commented out, than to maintain a complicated tree of feature branches that need to be rebased against one another. It also provides the reviewer some hints of my future intentions.