Further refactoring of Experiment class (#50)
Do more groundwork in preparation for supporting querying of time series and a metrics library.

Specifically, spin off the construction of the join conditions into a separate method so that get_per_client_data() fits on one screen and is easier to reason about.


* Move join condition construction to separate method.

* Trivial gardening

* Move data_source checking into data_source specific method.

In the future (with the metrics library) we will have multiple data_sources in one query, so that users can just list all the metrics they want to analyse without partitioning them by data_source. So all data_source functionality should eventually live inside a reduce loop - this makes that more possible.

* Rename `filter_*_for_analysis_window()` to `_process_*()`

In the future we are going to do significantly more than filtering in these methods. For now let's be less verbose, and make them private.

* Staticify static methods

* Push ugly aliases out of the way

* Lower the demanded precision from a random unit test

* Address review comments.