Change mind about analysis windows.
Previously, `TimeSeries.analysis_window_start` and
`analysis_window_end` were going to be:
* `int`s when querying non-time-series, or
* `Column`s when querying time series.

This was kind of a tortured overloading, and required using
global-style column construction in the `TimeSeries` constructor -
i.e. it had to reference a `Column` that gets constructed by
`_get_join_conditions()`. Ugh.

Instead, we set it to `None` and ignore it and in a future commit
we'll handle this logic in `_get_join_conditions()`.