Add `Experiment.add_time_series_to_enrollments()`
When querying time series, we need to query an extra dimension of
data (time!).

In order to obtain this extra dimension of data, there is a design
decision to be made with three choices:
1. Make more queries (e.g. one per time period/analysis window),
2. Switch from one row per client to one row per `(client, time)`, or
3. Switch from one column per metric to one column per `(metric, time)`

It would be good to avoid (1) because repeated querying is expensive
and the repeated queries would involve a bunch of overlapping data.

(2) seems more natural than (3): we can add an extra column to identify
the row's analysis window, then pivot in pandas later. If we were to
do (3) we would probably have to do some munging of column names.

In defence of (3): if we philosophically believe that a metric's
definition must involve its analysis window (e.g. "'retention' is a
sensible metric if defined as 'used the browser at all in week 3 after
enrolling'") then `time` in the `(metric, time)` tuple is redundant.

After briefly talking to @tdsmith some months ago I've gone with (2);
happy to revisit this if necessary.