Add time series (#51)
Add support for querying metrics as time series. Daily and weekly time series have official support in the `Experiment` class, though it's fairly easy to DIY an arbitrary set of analysis windows too.

This required the addition of new `TimeLimits` and `AnalysisWindow` classes, and a refactor of `Experiment`.

See https://github.com/mozilla/mozanalysis/pull/51 for the full gory details of the reasoning behind the design choices.