Refactor `Experiment._get_per_client_data()`, ready for batches.
We want users to be able to supply an arbitrary list of `Metric`s to `Experiment.get_per_client_data()` and `Experiment.get_time_series_data()`, without requiring that all metrics share a common `DataSource`. Therefore we need to handle joins with arbitrary numbers of `DataSource`s and associated `Metric`s.

This lays the groundwork for that: `Experiment._get_per_client_data()` now builds a `dict` of a list of metrics (expressed as Columns) associated with each data source (expressed as a DataFrame). At present this `dict` is specified manually and only has one key - but the downstream code is capable of handling `dict`s with arbitrary numbers of elements.

As such, this diff offers no new functionality; it merely lays the groundwork for the next steps.