Fully qualify result table names for TimeSeriesResult.
The data to be extracted from a TimeSeriesResult is identified by `(project_id, dataset_id, table_name)`. Previously, `project_id` and `dataset_id` were stored in a `BigQueryContext`, separate to the `table_name`. If someone was using multiple BigQueryContext instances, with different `project_id`s or `dataset_id`s (which is a terrible idea!) then they would need to correctly match each TimeSeriesResult to the right BigQueryContext.

To remove the chance for such mismatches, TimeSeriesResult now stores a fully qualified table name: `project_id.dataset_id.table_name`.

It would be nice to deal with fully qualified table names everywhere, but when specifying a destination table (when saving results into a table), the BigQuery API seems not to allow the destination table to be specified as a fully qualified table name - instead the project_id, dataset_id and table_name must be submitted separately in three separate places (respectively: when instantiating the bigquery client, when instantiating a dataset object, and when calling the `table()` method on that dataset object. Ergh.