Add `Experiment._build_segments_query()` and `_build_segments_query_bits()
This code is largely copypasta from `_build_metrics_query_bits()`.

The main difference is that this builds up a largely-self-contained query (it needs `raw_enrollments` to be defined), and this query will (in a later commit) be dropped into the master query in `Experiment.build_query()` between `raw_enrollments` and `enrollments`.

The strategy is the same: `SegmentDataSource.build_query()` (coming soon!) returns a SQL query that adds one column per segment from the data source while leaving the other columns untouched and not changing the number of rows. Then in `_build_segments_query()` we combine the new columns from each data source by LEFT JOINing repeatedly and SELECTing the new columns.