Add segments functionality (#76)
* Refactor code so we can reuse it for segments

* Add `Experiment._build_segments_query()` and `_build_segments_query_bits()

This code is largely copypasta from `_build_metrics_query_bits()`.

The main difference is that this builds up a largely-self-contained query (it needs `raw_enrollments` to be defined), and this query will (in a later commit) be dropped into the master query in `Experiment.build_query()` between `raw_enrollments` and `enrollments`.

The strategy is the same: `SegmentDataSource.build_query()` (coming soon!) returns a SQL query that adds one column per segment from the data source while leaving the other columns untouched and not changing the number of rows. Then in `_build_segments_query()` we combine the new columns from each data source by LEFT JOINing repeatedly and SELECTing the new columns.

* Add segment modules, tests, and docs.

Again we borrow ideas and code from the metrics library, making enough changes/simplifications that copypasta is better than DRY here.

`SegmentDataSource.build_query()` is the method that builds the segment queries that plug into `Experiment._build_segment_query()` as `LEFT JOIN`s.

Also adds three segments and their data source. These segments are documented in [this pull request to DTMO](https://github.com/mozilla/firefox-data-docs/pull/428). There is also a [related PR for bigquery-etl](https://github.com/mozilla/bigquery-etl/pull/825/); once that is merged it will be easier to query these segments: we would be able to replace the `select_expr` with the currently-commented-out versions, and replace the data source with a non-lagging window.

* Hook up segments functionality and expose it.

Let users specify a list of segments, and use the segments query in `Experiment.build_query()`.

* Trailing commas in SELECT are no longer considered malformed \o/

* Correct the `SegmentDataSource.build_query()` docstring.

And augment the `Experiment._build_segments_query()` docstring.

* Put `segment_list` as the final argument to avoid a breaking API change

* Move `sql_lint()` into a helpers module to avoid duplication

* Update segments to latest proposed definitions

Hopefully bigquery-etl will be updated before we need to merge this, so we'll be able to remove `tmp_segment_regular_users_v1` before merging.

* I miss fixtures and real tests :(

* Update usage regluarity segment to latest definition (v2)

* Update test for new variable name