COALESCE segments to remove NULL values (#79)
Because `clients_last_seen` is constructed from the `main` ping, and our enrollments are obtained from the `events` ping,there are enrollments without a `clients_last_seen` row that matches the submission date.

This only happens for users without any (main ping) activity in the past 28 days. Therefore these users belong in `new_irregular_users_v2`.

This problem will persist even once the segments are in the ETL.

There are terser solutions (e.g. `MAX(...=...) IS TRUE` for two segments to map `NULL` to `FALSE`), but I figured that this verbose form was the most readily understandable.