Fix the `days_of_use` metric. (#107)
@SuYoungHong noticed that the results of the `days_of_use` metric introduced in PR #105 were very suspicious.

The `COUNT(*)` was counting enrollments even when they had no row in `clients_daily`. Oops.

Instead, we need to `COUNT()` a non-null column from `clients_daily`; `COUNT(ds.submission_date)` seems to be the best for communicating what this metric does? And it's not worth adding a `DISTINCT` because it's implied by `clients_daily` and the lack of other joins?

I haven't tried this code yet but I've asked Su to do so in [this notebook](https://colab.research.google.com/drive/1E-f35o3DE6FIdnmNjs7eEA413lkNjtLh).