Fixes #117. Handle duplicate client_ids more correctly.
Metric DataSources were not specific-enough when joining to the `enrollments` view: when there are multiple rows for one client_id, with different branches, rows for all branches are used for all the enrollment rows. So a cloned profile in a three-branch experiment could have up to 3*7=21 days of use in a 7 day analysis window.

Really we want to `GROUP BY e.*` when aggregating over the data source, but being SQL we need to manually list the full set of columns that uniquely identify a row.

Then, when joining this result to the others in the relevant `Experiment` method, we need to use all these join keys again.