Handle duplicate client_ids more correctly.
Segment DataSources were not specific-enough when joining to the `enrollments` view: when there are multiple enrollments for one client_id (with different branches), the join won't be a one-to-many join from one enrollment to potentially many data source rows, it will be a many-to-many join, yielding potentially incorrect results due to the extra/duplicated datasource rows. If the enrollments have different enrollment_dates then there's further potential for error.

Really, we want to `GROUP BY e.*` when aggregating over the data source, but being SQL we need to manually list the full set of columns that uniquely identify a row.

Then, when using this in the relevant `Experiment` method, we need to list all these join keys again.