Fix `Experiment.get_single_window_data()`
It appears that #116 / edc44eb2 broke `Experiment.get_single_window_data()` when replacing `run_query()` with `run_query_and_fetch()`. This was not caught by our tests because we can't mock BQ, so we can't test that method in a meaningful way (i.e. if we had a mock for run_query() we may well have forgotten to delete the mock, leaving the test still to pass).

I'm updating the logic to match `Experiment.get_time_series_data()`, without a deeper understanding of what's going on.