Join enrollments more rigorously... and fix what was broken (#120)
* Fixes #117. Handle duplicate client_ids more correctly.

Metric DataSources were not specific-enough when joining to the `enrollments` view: when there are multiple rows for one client_id, with different branches, rows for all branches are used for all the enrollment rows. So a cloned profile in a three-branch experiment could have up to 3*7=21 days of use in a 7 day analysis window.

Really we want to `GROUP BY e.*` when aggregating over the data source, but being SQL we need to manually list the full set of columns that uniquely identify a row.

Then, when joining this result to the others in the relevant `Experiment` method, we need to use all these join keys again.

* Handle duplicate client_ids more correctly.

Segment DataSources were not specific-enough when joining to the `enrollments` view: when there are multiple enrollments for one client_id (with different branches), the join won't be a one-to-many join from one enrollment to potentially many data source rows, it will be a many-to-many join, yielding potentially incorrect results due to the extra/duplicated datasource rows. If the enrollments have different enrollment_dates then there's further potential for error.

Really, we want to `GROUP BY e.*` when aggregating over the data source, but being SQL we need to manually list the full set of columns that uniquely identify a row.

Then, when using this in the relevant `Experiment` method, we need to list all these join keys again.

* Fix `Experiment.get_single_window_data()`

It appears that #116 / edc44eb2 broke `Experiment.get_single_window_data()` when replacing `run_query()` with `run_query_and_fetch()`. This was not caught by our tests because we can't mock BQ, so we can't test that method in a meaningful way (i.e. if we had a mock for run_query() we may well have forgotten to delete the mock, leaving the test still to pass).

I'm updating the logic to match `Experiment.get_time_series_data()`, without a deeper understanding of what's going on.

* Fix get_single_window_data() and get_time_series_data().

It appears that the method argument `enrollments_table` was misinterpreted as requiring a BigQuery results object rather than the name of the enrollments table, as a string.

* Fix BigQueryContext.run_query_and_fetch().

In some contexts, we need to provide unqualified table names. In others, it must be fully qualified.

* Don't enquote the fully qualified table name, and hope for the best.

bigquery.Client().list_rows() is unhappy when passed a fully qualified table name enquoted with backticks. So we'll pass it a nude one. Hopefully this doesn't break all our other queries...?

* Fix build_metrics_query().

The `metrics_joins` rely on access to a table named `enrollments`. Provide it!

Also be more honest in the docs for build_enrollments_query().

I changed the alias for segmented_enrollments from `e` to `se` to make it clearer that segment information is in there too.

* Address review comments

* Rename run_query

* Document requirement for custom_enrollments_query

* Go back to a trivial CTE

since table aliases apparently aren't respected inside join subqueries.

* Make a print statement more ambiguous

It could be printed for the enrollments table, not just the results.