Use concrete table for main data source
Uses the `telemetry_stable.main_v4` table instead of the `telemetry.main` compatibility view for the `main` data source.

This is intended to reduce the complexity of the analysis query in order to stay within Google's complexity budget.

We expect that this is compatible with all existing metrics because the view doesn't make very many changes, although the changes it does make are fairly complex.