Improve the performance of bayesian bootstrap by 10x
The vast majority of the current time is in:
np.random.RandomState(unique_seed)

Previously, when generation of the sample weights was parallelized it
made sense to reseed for each sampling so that it was
deterministic. However, that parallelization was removed in
e33914d7fa42bfa1f305d14ee7b96c3a39643abb

With that gone we can drop the reseeding entirely which makes
16 or so 100 sample calls to compare_branches go from 50seconds
down to 5.